requirements.txt
africastalking==2.0.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.4.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cron_descriptor==2.0.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.6
django-auditlog==3.4.1
django-axes==8.3.1
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fonttools==4.61.1
gunicorn==23.0.0
h11==0.16.0
humanize==4.13.0
idna==3.10
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
numpy==2.4.2
openpyxl==3.1.5
packaging==25.0
pandas==3.0.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.12.1
PyJWT==2.10.1
pyotp==2.9.0
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
referencing==0.37.0
requests==2.32.5
responses==0.26.0
rpds-py==0.30.0
schema==0.7.8
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
